Refactor app into modular OOP structure
